Farazbranch (#3)
* converted .py file from the pop.ipynb file 

This is a .py file for arriving at the final population dataset which will be merged with our other two datasets for the analysis. The main branch contains the .ipynb file, I just added this .py file here for reference. Pls check and accept my pull request tnx

* py files for mortality and shipment

* After your comments

Added checks and removed any errors

* update folders

* update folders

* update folder name

* adding source data for pop

* Added the source datasets downloaded from nhgis

Contains the raw csv files and also a pop.ipynb file which explains how we used these files and what steps were taken to come to a complete and final dataset which is in this directory with the name "final_population_data.csv"

* Delete 00_Data_Cleaning_and_Merging/01_Source_Data/population/xyz

* Create folder for shipment data

* Data and notebook file for shipment data

chunk.ipynb is the file for processing
`final_shipments.csv` is the file after processing and chunking
'shipments_data' is the parquet file

* cleaning up

* Delete 00_Data_Cleaning_and_Merging/01_Source_Data/shipments/xyz

* Create mortality folder

* All files

The .txt files are the raw data, i processed it using the ipynb file and compiled everything in the mortality.csv file

* update

* update code draft

* update code draft

* update pop code draft

* Add files via upload

* approve .py merge request

* did some data merging

* merging

* completed .py file

* merged csv created and comments resolved

* updated .py file

* removed conflicting files

---------

Co-authored-by: tinayiluo0322 <143360909+tinayiluo0322@users.noreply.github.com>